refactor-composables: move DisposableInput and CustomInput to widget
